requirements.txt
aiogram==3.*
vk-api
python-dotenv
requests
